moves 'IndicadorPaginaActual' into a new component
